requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
numpy==2.2.6
packaging==25.0
ping3==4.0.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-nmap==0.7.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.25.1
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
